Add @CRUD.BeforeObjectSave annotation for intercepting object.save called by CRUD.create/save.
Methods annotated with @BeforeObjectSave must accept the Model object as argument and
return a boolean, that indicates if the calling method (create/save) shall save the object.
This is useful if object creation/saving is more expensive and shall be done in the background,
i.e. delegated to a Job.